sign in with google is done

+++ controllers/api/v1/posts_api.js
@@ -19,29 +19,30 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    if (req.xhr) {
-      return res.status(200).json({
-        data: {
-          post_id: req.params.id,
-        },
-        message: "Post deleted",
+      if (req.xhr) {
+        return res.status(200).json({
+          data: {
+            post_id: req.params.id,
+          },
+          message: "Post deleted",
+        });
+      }
+
+      req.flash("success", "Post and associated comments deleted!");
+
+      return res.json(200, {
+        message: "Post and associated comments delete successfully",
+      });
+    } else {
+      res.json(401, {
+        message: "You cannot delete post",
       });
     }
-
-    req.flash("success", "Post and associated comments deleted!");
-
-    return res.json(200, {
-      message: "Post and associated comments delete successfully",
-    });
-    // } else {
-    //   req.flash("error", "You cannot delete this post!");
-    //   return res.redirect("back");
-    // }
   } catch (err) {
     return res.json(500, {
       message: "Internal Server Error",
